chess-player-agent/pyproject.toml
[project]
name = "chess-player-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai-slim[a2a,mcp,openai]>=0.7.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
]
